requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
Deprecated==1.2.12
github3.py==2.0.0
idna==3.2
jwcrypto==0.9.1
pycparser==2.20
PyGithub==1.55
pygithub3==0.5.1
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
